move precalculation of density integral to interactionRate.py
