+++ interactionRate.py
@@ -1,5 +1,6 @@
 import numpy as np
 from scipy.integrate import cumulative_trapezoid, romb, quad
+import os
 
 eV = 1.60217657e-19  # [J]
 Mpc = 3.08567758e22  # [m]
@@ -50,6 +51,8 @@ def calc_rate_s(s_kin, xs, E, field, z=0, cdf=False):
     """
 
     if cdf:
+        # precalculate the field integral if it not exisist and load it afterwards
+        calculateDensityIntegral(field)
         file = "data/fieldDensity/" + field.name + ".txt"
         densityIntegral = np.loadtxt(file)
 
@@ -67,6 +70,44 @@ def calc_rate_s(s_kin, xs, E, field, z=0, cdf=False):
         y = n * F / s_kin
         ds = mean_log_spacing(s_kin)
         return romb(y, dx=ds) / 2 / E * Mpc
+
+def calculateDensityIntegral(field):
+    """ Precalculate the integral over the density 
+        int_{Emin}^{Emax} n(eps) / eps^2  deps 
+        and save as a file.
+
+        field : photon background, see photonField.py
+    """
+
+    # precalc the photon density integral 
+    Emax = field.getEmax()
+    Emin =  1e4 / 4 / 1e23 * eV # min(s_kin) / 4 / max(E_e)
+    alpha = np.logspace(np.log10(Emin), np.log10(Emax), 10000) # lower boundary of the integral.
+
+    # check if file already exist
+    folder = "data/fieldDensity/"
+    if not os.path.isdir(folder):
+        os.makedirs(folder)
+    file = folder + field.name + ".txt"
+    if os.path.isfile(file):
+        return # file already existst no calculation necessary
+
+    # calculate integral
+    I_gamma = np.zeros_like(alpha)
+    for i in range(len(alpha)):
+        I_gamma[i] = quad(lambda E: field.getDensity(E) / E**2, a = alpha[i], b = Emax, full_output=1)[0]
+
+    # save file
+    header = "# calculate integral n(e)/e^2 de from eMin to eMax, where eMax is the maximal photon energy of the background \n"
+    try: 
+        git_hash = gh.get_git_revision_hash()
+        header += "Produced with crpropa-data version: "+git_hash+"\n"
+        header += "# eMin [eV]\tintegral\n"
+    except:
+        header += "# eMin [eV]\tintegral\n"
+    data = np.c_[alpha, I_gamma]
+    fmt = '%.4e\t%8.7e'
+    np.savetxt(file, data, fmt = fmt, header = header)
 
 
 def mean_log_spacing(x):
